src/magnetar_deer/agent/playgound.test.py
import os
from dotenv import load_dotenv

from google import genai

load_dotenv("config.env")
api_key = os.getenv("GEMINI_API_KEY")

# The client gets the API key from the environment variable `GEMINI_API_KEY`.
client = genai.Client()

response = client.models.generate_content(
    model="gemini-3-flash-preview", contents="Explain how AI works in a few words"
)
print(response.text)